Move screen stuff to MC module
