+++ minecraft/archCommon/src/main/java/io/github/kosmx/emotes/arch/gui/EmoteMenuImpl.java
@@ -3,10 +3,10 @@ package io.github.kosmx.emotes.arch.gui;
 import io.github.kosmx.emotes.arch.gui.screen.*;
 import io.github.kosmx.emotes.arch.gui.widgets.AbstractEmoteListWidget;
 import io.github.kosmx.emotes.main.EmoteHolder;
-import io.github.kosmx.emotes.main.screen.AbstractScreenLogic;
-import io.github.kosmx.emotes.main.screen.EmoteMenu;
-import io.github.kosmx.emotes.main.screen.IScreenSlave;
-import io.github.kosmx.emotes.main.screen.widget.IEmoteListWidgetHelper;
+import io.github.kosmx.emotes.arch.screen.AbstractScreenLogic;
+import io.github.kosmx.emotes.arch.screen.EmoteMenu;
+import io.github.kosmx.emotes.arch.screen.IScreenSlave;
+import io.github.kosmx.emotes.arch.screen.widget.IEmoteListWidgetHelper;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.events.GuiEventListener;
