+++ minecraft/archCommon/src/main/java/io/github/kosmx/emotes/arch/gui/screen/AbstractControlledModScreen.java
@@ -1,6 +1,5 @@
 package io.github.kosmx.emotes.arch.gui.screen;
 
-import com.mojang.blaze3d.vertex.PoseStack;
 import io.github.kosmx.emotes.arch.executor.types.TextImpl;
 import io.github.kosmx.emotes.executor.EmoteInstance;
 import io.github.kosmx.emotes.executor.dataTypes.Text;
@@ -9,9 +8,9 @@ import io.github.kosmx.emotes.executor.dataTypes.screen.IScreen;
 import io.github.kosmx.emotes.executor.dataTypes.screen.widgets.IButton;
 import io.github.kosmx.emotes.executor.dataTypes.screen.widgets.ITextInputWidget;
 import io.github.kosmx.emotes.executor.dataTypes.screen.widgets.IWidget;
-import io.github.kosmx.emotes.main.screen.AbstractScreenLogic;
-import io.github.kosmx.emotes.main.screen.IScreenLogicHelper;
-import io.github.kosmx.emotes.main.screen.IScreenSlave;
+import io.github.kosmx.emotes.arch.screen.AbstractScreenLogic;
+import io.github.kosmx.emotes.arch.screen.IScreenLogicHelper;
+import io.github.kosmx.emotes.arch.screen.IScreenSlave;
 import net.minecraft.Util;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
