+++ minecraft/archCommon/src/main/java/io/github/kosmx/emotes/arch/gui/screen/IDrawableImpl.java
@@ -1,12 +1,11 @@
 package io.github.kosmx.emotes.arch.gui.screen;
 
 import com.mojang.blaze3d.systems.RenderSystem;
-import com.mojang.blaze3d.vertex.PoseStack;
 import io.github.kosmx.emotes.arch.executor.types.IdentifierImpl;
 import io.github.kosmx.emotes.arch.executor.types.TextImpl;
 import io.github.kosmx.emotes.executor.dataTypes.IIdentifier;
 import io.github.kosmx.emotes.executor.dataTypes.Text;
-import io.github.kosmx.emotes.main.screen.IRenderHelper;
+import io.github.kosmx.emotes.arch.screen.IRenderHelper;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.renderer.GameRenderer;
