+++ minecraft/archCommon/src/main/java/io/github/kosmx/emotes/arch/gui/screen/ingame/FullScreenListImpl.java
@@ -1,16 +1,15 @@
 package io.github.kosmx.emotes.arch.gui.screen.ingame;
 
-import com.mojang.blaze3d.vertex.PoseStack;
 import io.github.kosmx.emotes.arch.gui.EmoteMenuImpl;
 import io.github.kosmx.emotes.arch.gui.screen.AbstractControlledModScreen;
 import io.github.kosmx.emotes.arch.gui.widgets.AbstractEmoteListWidget;
 import io.github.kosmx.emotes.executor.dataTypes.screen.IScreen;
 import io.github.kosmx.emotes.main.EmoteHolder;
 import io.github.kosmx.emotes.main.network.ClientEmotePlay;
-import io.github.kosmx.emotes.main.screen.AbstractScreenLogic;
-import io.github.kosmx.emotes.main.screen.IScreenSlave;
-import io.github.kosmx.emotes.main.screen.ingame.FullMenuScreenHelper;
-import io.github.kosmx.emotes.main.screen.widget.IEmoteListWidgetHelper;
+import io.github.kosmx.emotes.arch.screen.AbstractScreenLogic;
+import io.github.kosmx.emotes.arch.screen.IScreenSlave;
+import io.github.kosmx.emotes.arch.screen.ingame.FullMenuScreenHelper;
+import io.github.kosmx.emotes.arch.screen.widget.IEmoteListWidgetHelper;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.events.GuiEventListener;
