+++ minecraft/archCommon/src/main/java/io/github/kosmx/emotes/arch/screen/AbstractScreenLogic.java
@@ -1,4 +1,4 @@
-package io.github.kosmx.emotes.main.screen;
+package io.github.kosmx.emotes.arch.screen;
 
 import java.nio.file.Path;
 import java.util.List;
