+++ minecraft/archCommon/src/main/java/io/github/kosmx/emotes/arch/screen/IScreenLogicHelper.java
@@ -1,4 +1,4 @@
-package io.github.kosmx.emotes.main.screen;
+package io.github.kosmx.emotes.arch.screen;
 
 import io.github.kosmx.emotes.executor.dataTypes.Text;
 import io.github.kosmx.emotes.executor.dataTypes.screen.IConfirmScreen;
